+++ taf/tools/repo/__init__.py
@@ -10,6 +10,7 @@ from taf.log import initialize_logger_handlers, taf_logger
 from taf.tools.cli import catch_cli_exception, find_repository
 from taf.updater.types.update import UpdateType
 from taf.updater.updater import OperationType, UpdateConfig, clone_repository, update_repository, validate_repository
+from taf.yubikey.yubikey_manager import pin_managed
 
 
 def common_update_options(f):
@@ -24,7 +25,7 @@ def common_update_options(f):
 
 def _call_updater(config, format_output):
     """
-    A helper function which call update or clone repository
+    A helper function which calls update or clone repository
     """
     try:
         if config.operation == OperationType.CLONE:
@@ -120,9 +121,11 @@ def create_repo_command():
                   "committed automatically")
     @click.option("--test", is_flag=True, default=False, help="Indicates if the created repository "
                   "is a test authentication repository")
-    def create(path, keys_description, keystore, no_commit, test):
+    @pin_managed
+    def create(path, keys_description, keystore, no_commit, test, pin_manager):
         create_repository(
             path=path,
+            pin_manager=pin_manager,
             keystore=keystore,
             roles_key_infos=keys_description,
             commit=not no_commit,